refactor: move file from atom to components as atom doesn't exist in new code structure
