+++ src/features/home/atoms/index.tsx
@@ -1,4 +1,3 @@
-import { HomeOfferTile } from './HomeOfferTile'
 import { ModuleTitle } from './ModuleTitle'
 import { SeeMore } from './SeeMore'
-export { HomeOfferTile, ModuleTitle, SeeMore }
+export { ModuleTitle, SeeMore }
