+++ src/features/home/components/modules/OffersModule.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect } from 'react'
 
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useOfferModule } from 'features/home/api/useOfferModule'
-import { HomeOfferTile, SeeMore } from 'features/home/atoms'
+import { SeeMore } from 'features/home/atoms'
 import {
   SearchParametersFields,
   DisplayParametersFields,
@@ -19,6 +19,7 @@ import { SearchHit, useParseSearchParameters } from 'libs/search'
 import { useCategoryIdMapping, useCategoryHomeLabelMapping } from 'libs/subcategories'
 import { PassPlaylist } from 'ui/components/PassPlaylist'
 import { CustomListRenderItem, RenderFooterItem } from 'ui/components/Playlist'
+import { HomeOfferTile } from 'features/home/components/HomeOfferTile'
 
 type OffersModuleProps = {
   search: SearchParametersFields[]
